# Conflicts:
#	src_main/engine/qcommon.h
#	src_main/engine/sdl/sys_sdl.c

+++ src_main/utils/qe4/cmdlib.c
@@ -40,7 +40,7 @@ double I_FloatTime (void)
 
 	time (&t);
 
-	return t;
+	return (double)t;
 #if 0
 // more precise, less portable
 	struct timeval tp;
